requirements.txt
PySide6==6.5.1
PyYAML==6.0
